Rename factorChallengeTests
